Set character mode for input stream
