+++ pkg/watch/key_watcher_unix.go
@@ -1,0 +1,21 @@
+//go:build aix || darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris || zos
+// +build aix darwin dragonfly freebsd linux netbsd openbsd solaris zos
+
+package watch
+
+import (
+	"golang.org/x/sys/unix"
+)
+
+func enableCharInput(fd int) error {
+	termios, err := unix.IoctlGetTermios(fd, ioctlReadTermios)
+	if err != nil {
+		return err
+	}
+	termios.Lflag &^= unix.ICANON
+	if err := unix.IoctlSetTermios(fd, ioctlWriteTermios, termios); err != nil {
+		return err
+	}
+
+	return nil
+}
